fix(serialization): incorrect map serialization
When switching to Symfony's serializer, the MapDenormalizer was implemented, but not added to the Serializer parameters.
Thus, the attributes of a realm and other fields using the Map type were always empty.

Additionally, the MapNormalizerTest contained test data for the denormalizer. We adjusted this to correctly ingest Map's as the test values.

To properly normalize the Map type, we made the inner array publicly accessible. Previously the ArrayObject would otherwise hold an array with the entire Map object wrapped by the Map classname. This was addressed, by directly accessing the inner array and passing it to the ArrayObject.

We also added an integration test to prevent these kind of issues in the future, and to test the update method for realm attributes.

(cherry picked from commit b5bffe807c02d106ef47a3e27c3568ffda0f2847)